Cell-cycle markers

Cell-cycle markers

CellCycleMarkers(object, gene2symbol)

importCellCycleMarkers(file, organism, release)

Arguments

object
Object.

gene2symbol
Gene2Symbol. Gene-to-symbol mappings. Must contain geneID and geneName columns. See Gene2Symbol for more information.

file
character(1). File path.

organism
character(1). Full Latin organism name (e.g. "Homo sapiens").

release
integer(1). Ensembl release version (e.g. 90). We recommend setting this value if possible, for improved reproducibility. When left unset, the latest release available via AnnotationHub/ensembldb is used. Note that the latest version available can vary, depending on the versions of AnnotationHub and ensembldb in use.

Value

CellCycleMarkers.

Note

Updated 2020-10-12.

Examples

markers_dir <- system.file(
    file.path("extdata", "markers"),
    package = "pointillism"
)

cell_cycle_dir <- file.path(markers_dir, "cell-cycle")
files <- list.files(cell_cycle_dir, pattern = "*.csv", full.names = TRUE)
file <- files[[1L]]

## Ensembl release version.
release_file <- file.path(markers_dir, "ensembl-release.txt")
release <- as.integer(readLines(release_file))

importCellCycleMarkers(
    file = file,
    organism = basenameSansExt(file),
    release = release
)

#> → Importing drosophila-melanogaster.csv at /private/var/folders/l1/8y8sjzmn15v49jgrqglghcfr0000gn/T/RtmpzzDSHS/temp_libpath6e6b4c6fe7f2/pointillism/extdata/markers/cell-cycle using vroom::`vroom()`.

#> → Making `GRanges` from Ensembl.

#> → Getting `EnsDb` from AnnotationHub 2.20.2 (2020-04-27).

#> Error in .getAnnotationHubID(organism = organism, genomeBuild = genomeBuild,     release = release): Assert failure.
#> [6] all(mcols[["species"]] == organism) is not TRUE.