Cell-type markers

Cell-type markers

CellTypeMarkers(object, gene2symbol)

importCellTypeMarkers(file, organism, release)

Arguments

object
Object.

gene2symbol
Gene2Symbol. Gene-to-symbol mappings. Must contain geneID and geneName columns. See Gene2Symbol for more information.

file
character(1). File path.

organism
character(1). Full Latin organism name (e.g. "Homo sapiens").

release
integer(1). Ensembl release version (e.g. 90). We recommend setting this value if possible, for improved reproducibility. When left unset, the latest release available via AnnotationHub/ensembldb is used. Note that the latest version available can vary, depending on the versions of AnnotationHub and ensembldb in use.

Value

CellTypeMarkers.

Note

Updated 2020-02-21.

Examples

markers_dir <- system.file(
    file.path("extdata", "markers"),
    package = "pointillism"
)

cell_type_dir <- file.path(markers_dir, "cell-type")
files <- list.files(cell_type_dir, pattern = "*.csv", full.names = TRUE)
file <- files[[1L]]

## Ensembl release version.
release_file <- file.path(markers_dir, "ensembl-release.txt")
release <- as.integer(readLines(release_file))

importCellTypeMarkers(
    file = file,
   organism = basenameSansExt(file),
    release = release
)

#> → Importing drosophila-melanogaster.csv using vroom::`vroom()`.

#> → Making `GRanges` from Ensembl.

#> → Matching `EnsDb` from AnnotationHub 2.20.0 (2020-04-27).

#> ℹ AH78755: Ensembl 99 EnsDb for Drosophila melanogaster.

#> → Making `GRanges` from `EnsDb`.

#>     Organism: Drosophila melanogaster

#>     Genome build: BDGP6.28

#>     Ensembl release: 99

#>     Level: genes

#> → Defining `broadClass` using: `geneBiotype, geneName, seqnames`.

#> → Arranging by `geneID`.

#> ℹ 17807 genes detected.

#> CellTypeMarkers
#> Drosophila melanogaster (Ensembl 99)
#> crystal_cell(5): klu lz ... PPO1 Sox102F
#> lamellocyte(5): 14-3-3zeta atilla ... pnt tx
#> plasmatocyte(5): bbx bon ... Sox102F ttk
#> prohemocyte(5): dome h ... Tep4 wg