Sanitize Seurat Markers

This generator function is designed to take the original return from a Seurat marker analysis and add corresponding gene annotations.

SeuratMarkers(markers, ranges, alpha = 0.05)

Arguments

markers
data.frame. Unmodified Seurat::FindAllMarkers() or Seurat::FindMarkers() return.

ranges
GRanges. Gene annotations. Names must correspond to the rownames defined in seurat@data. The function will automatically subset the ranges and arrange them alphabetically.

alpha
scalar numeric. Alpha cutoff (adjusted P value; false discovery rate).

Value

SeuratMarkers. Results are arranged by adjusted P value, and grouped per cluster if applicable.

Note

Seurat::FindAllMarkers() maps the counts matrix rownames correctly in the gene column, whereas Seurat::FindMarkers() maps them correctly in the rownames of the returned marker data.frame.

Examples

data(seurat_small)
object <- seurat_small

## `FindAllMarkers()` return.
invisible(capture.output(
    all_markers <- Seurat::FindAllMarkers(object)
))
all_sanitized <- SeuratMarkers(
    markers = all_markers,
    ranges = rowRanges(object)
)

#> `Seurat::FindAllMarkers()` return detected

str(all_sanitized)

#> Formal class 'SeuratMarkers' [package "pointillism"] with 6 slots
#>   ..@ rownames       : NULL
#>   ..@ nrows          : int 208
#>   ..@ listData       :List of 8
#>   .. ..$ cluster : Factor w/ 4 levels "0","1","2","3": 1 1 1 1 1 1 1 1 1 1 ...
#>   .. ..$ name    : chr [1:208] "LTB" "TMEM123" "CXCR4" "SRSF7" ...
#>   .. ..$ avgLogFC: num [1:208] 2.5 2.09 1.41 1.71 3.56 ...
#>   .. ..$ pct1    : num [1:208] 1 0.69 0.759 0.759 0.379 0.552 0.034 0.586 0.724 0.207 ...
#>   .. ..$ pct2    : num [1:208] 0.333 0.216 0.275 0.333 0.02 0.137 0.529 0.196 0.392 0.765 ...
#>   .. ..$ pvalue  : num [1:208] 1.54e-11 2.16e-06 3.94e-06 1.14e-05 1.47e-05 ...
#>   .. ..$ padj    : num [1:208] 3.54e-09 4.96e-04 9.05e-04 2.61e-03 3.37e-03 ...
#>   .. ..$ ranges  :Formal class 'GRanges' [package "GenomicRanges"] with 7 slots
#>   .. .. .. ..@ seqnames       :Formal class 'Rle' [package "S4Vectors"] with 4 slots
#>   .. .. .. .. .. ..@ values         : Factor w/ 273 levels "X","20","1","6",..: 28 8 11 19 8 10 18 5 3 8 ...
#>   .. .. .. .. .. ..@ lengths        : int [1:196] 1 1 2 1 1 1 1 1 1 2 ...
#>   .. .. .. .. .. ..@ elementMetadata: NULL
#>   .. .. .. .. .. ..@ metadata       : list()
#>   .. .. .. ..@ ranges         :Formal class 'IRanges' [package "IRanges"] with 6 slots
#>   .. .. .. .. .. ..@ start          : int [1:208] 31535744 102267063 136871919 38970741 31677579 118209669 79825597 35852797 186500994 161185024 ...
#>   .. .. .. .. .. ..@ width          : int [1:208] 1968 74053 3817 7896 10942 3791 3686 26909 6696 5466 ...
#>   .. .. .. .. .. ..@ NAMES          : chr [1:208] "LTB" "TMEM123" "CXCR4" "SRSF7" ...
#>   .. .. .. .. .. ..@ elementType    : chr "ANY"
#>   .. .. .. .. .. ..@ elementMetadata: NULL
#>   .. .. .. .. .. ..@ metadata       : list()
#>   .. .. .. ..@ strand         :Formal class 'Rle' [package "S4Vectors"] with 4 slots
#>   .. .. .. .. .. ..@ values         : Factor w/ 3 levels "+","-","*": 2 1 2 1 2 1 2 1 2 1 ...
#>   .. .. .. .. .. ..@ lengths        : int [1:97] 7 5 3 4 7 2 7 1 1 2 ...
#>   .. .. .. .. .. ..@ elementMetadata: NULL
#>   .. .. .. .. .. ..@ metadata       : list()
#>   .. .. .. ..@ seqinfo        :Formal class 'Seqinfo' [package "GenomeInfoDb"] with 4 slots
#>   .. .. .. .. .. ..@ seqnames   : chr [1:273] "X" "20" "1" "6" ...
#>   .. .. .. .. .. ..@ seqlengths : int [1:273] 155270560 63025520 249250621 171115067 198022430 159138663 133851895 135006516 191154276 81195210 ...
#>   .. .. .. .. .. ..@ is_circular: logi [1:273] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   .. .. .. .. .. ..@ genome     : chr [1:273] "GRCh37" "GRCh37" "GRCh37" "GRCh37" ...
#>   .. .. .. ..@ elementMetadata:Formal class 'DataFrame' [package "S4Vectors"] with 6 slots
#>   .. .. .. .. .. ..@ rownames       : NULL
#>   .. .. .. .. .. ..@ nrows          : int 208
#>   .. .. .. .. .. ..@ listData       :List of 10
#>   .. .. .. .. .. .. ..$ broadClass            : Factor w/ 8 levels "coding","ig",..: 1 1 1 1 1 1 1 1 1 1 ...
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr [1:208] "ENSG00000204487" "ENSG00000152558" "ENSG00000121966" "ENSG00000115875" ...
#>   .. .. .. .. .. .. ..$ entrezID              :List of 208
#>   .. .. .. .. .. .. .. ..$ ENSG00000204487: int 4050
#>   .. .. .. .. .. .. .. ..$ ENSG00000152558: int 114908
#>   .. .. .. .. .. .. .. ..$ ENSG00000121966: int 7852
#>   .. .. .. .. .. .. .. ..$ ENSG00000115875: int 6432
#>   .. .. .. .. .. .. .. ..$ ENSG00000100100: int 113791
#>   .. .. .. .. .. .. .. ..$ ENSG00000167286: int 915
#>   .. .. .. .. .. .. .. ..$ ENSG00000141522: int 396
#>   .. .. .. .. .. .. .. ..$ ENSG00000168685: int 3575
#>   .. .. .. .. .. .. .. ..$ ENSG00000156976: int [1:6] 100302143 677847 619568 619567 6043 1974
#>   .. .. .. .. .. .. .. ..$ ENSG00000158869: int 2207
#>   .. .. .. .. .. .. .. ..$ ENSG00000084207: int 2950
#>   .. .. .. .. .. .. .. ..$ ENSG00000198851: int 916
#>   .. .. .. .. .. .. .. ..$ ENSG00000101439: int 1471
#>   .. .. .. .. .. .. .. ..$ ENSG00000164691: int 117289
#>   .. .. .. .. .. .. .. ..$ ENSG00000126353: int 1236
#>   .. .. .. .. .. .. .. ..$ ENSG00000172005: int 4118
#>   .. .. .. .. .. .. .. ..$ ENSG00000177156: int 6888
#>   .. .. .. .. .. .. .. ..$ ENSG00000167460: int 7171
#>   .. .. .. .. .. .. .. ..$ ENSG00000100097: int 3956
#>   .. .. .. .. .. .. .. ..$ ENSG00000145220: int 55646
#>   .. .. .. .. .. .. .. ..$ ENSG00000170458: int 929
#>   .. .. .. .. .. .. .. ..$ ENSG00000137078: int 27240
#>   .. .. .. .. .. .. .. ..$ ENSG00000130254: int 9667
#>   .. .. .. .. .. .. .. ..$ ENSG00000104921: int 2208
#>   .. .. .. .. .. .. .. ..$ ENSG00000025708: int 1890
#>   .. .. .. .. .. .. .. ..$ ENSG00000169813: int 3185
#>   .. .. .. .. .. .. .. ..$ ENSG00000136732: int 2995
#>   .. .. .. .. .. .. .. ..$ ENSG00000072818: int 9744
#>   .. .. .. .. .. .. .. ..$ ENSG00000163453: int 3490
#>   .. .. .. .. .. .. .. ..$ ENSG00000100453: int 3002
#>   .. .. .. .. .. .. .. ..$ ENSG00000164733: int 1508
#>   .. .. .. .. .. .. .. ..$ ENSG00000011600: int 7305
#>   .. .. .. .. .. .. .. ..$ ENSG00000143546: int 6279
#>   .. .. .. .. .. .. .. ..$ ENSG00000163479: int 6746
#>   .. .. .. .. .. .. .. ..$ ENSG00000080824: int 3320
#>   .. .. .. .. .. .. .. ..$ ENSG00000129277: int 6351
#>   .. .. .. .. .. .. .. ..$ ENSG00000168497: int 8436
#>   .. .. .. .. .. .. .. ..$ ENSG00000163220: int 6280
#>   .. .. .. .. .. .. .. ..$ ENSG00000081059: int 6932
#>   .. .. .. .. .. .. .. ..$ ENSG00000151500: int 29087
#>   .. .. .. .. .. .. .. ..$ ENSG00000111716: int 3945
#>   .. .. .. .. .. .. .. ..$ ENSG00000233276: int 2876
#>   .. .. .. .. .. .. .. ..$ ENSG00000185201: int 10581
#>   .. .. .. .. .. .. .. ..$ ENSG00000025708: int 1890
#>   .. .. .. .. .. .. .. ..$ ENSG00000090382: int 4069
#>   .. .. .. .. .. .. .. ..$ ENSG00000101439: int 1471
#>   .. .. .. .. .. .. .. ..$ ENSG00000170458: int 929
#>   .. .. .. .. .. .. .. ..$ ENSG00000084207: int 2950
#>   .. .. .. .. .. .. .. ..$ ENSG00000030582: int 2896
#>   .. .. .. .. .. .. .. ..$ ENSG00000204482: int 7940
#>   .. .. .. .. .. .. .. ..$ ENSG00000126759: int 5199
#>   .. .. .. .. .. .. .. ..$ ENSG00000110077: int 64231
#>   .. .. .. .. .. .. .. ..$ ENSG00000163220: int 6280
#>   .. .. .. .. .. .. .. ..$ ENSG00000215048: int 3115
#>   .. .. .. .. .. .. .. ..$ ENSG00000113269: int 55819
#>   .. .. .. .. .. .. .. ..$ ENSG00000161570: int 6352
#>   .. .. .. .. .. .. .. ..$ ENSG00000168384: int 3113
#>   .. .. .. .. .. .. .. ..$ ENSG00000204472: int 199
#>   .. .. .. .. .. .. .. ..$ ENSG00000141505: int 432
#>   .. .. .. .. .. .. .. ..$ ENSG00000197746: int 5660
#>   .. .. .. .. .. .. .. ..$ ENSG00000143546: int 6279
#>   .. .. .. .. .. .. .. ..$ ENSG00000100079: int 3957
#>   .. .. .. .. .. .. .. ..$ ENSG00000100097: int 3956
#>   .. .. .. .. .. .. .. ..$ ENSG00000015475: int 637
#>   .. .. .. .. .. .. .. ..$ ENSG00000142089: int 10410
#>   .. .. .. .. .. .. .. ..$ ENSG00000162444: int 116362
#>   .. .. .. .. .. .. .. ..$ ENSG00000104870: int 2217
#>   .. .. .. .. .. .. .. ..$ ENSG00000179639: int 2205
#>   .. .. .. .. .. .. .. ..$ ENSG00000085265: int 2219
#>   .. .. .. .. .. .. .. ..$ ENSG00000204397: int 114769
#>   .. .. .. .. .. .. .. ..$ ENSG00000204287: int 3122
#>   .. .. .. .. .. .. .. ..$ ENSG00000163191: int 6282
#>   .. .. .. .. .. .. .. ..$ ENSG00000100906: int 4792
#>   .. .. .. .. .. .. .. ..$ ENSG00000116741: int 5997
#>   .. .. .. .. .. .. .. ..$ ENSG00000196126: int [1:4] 100996809 100507714 100507709 3123
#>   .. .. .. .. .. .. .. ..$ ENSG00000158869: int 2207
#>   .. .. .. .. .. .. .. ..$ ENSG00000197249: int 5265
#>   .. .. .. .. .. .. .. ..$ ENSG00000197405: int 728
#>   .. .. .. .. .. .. .. ..$ ENSG00000216490: int 10437
#>   .. .. .. .. .. .. .. ..$ ENSG00000145220: int 55646
#>   .. .. .. .. .. .. .. ..$ ENSG00000226264: int 3109
#>   .. .. .. .. .. .. .. ..$ ENSG00000204257: int 3108
#>   .. .. .. .. .. .. .. ..$ ENSG00000164733: int 1508
#>   .. .. .. .. .. .. .. ..$ ENSG00000173762: int 924
#>   .. .. .. .. .. .. .. ..$ ENSG00000101336: int 3055
#>   .. .. .. .. .. .. .. ..$ ENSG00000198502: int 3127
#>   .. .. .. .. .. .. .. ..$ ENSG00000011600: int 7305
#>   .. .. .. .. .. .. .. ..$ ENSG00000197766: int 1675
#>   .. .. .. .. .. .. .. ..$ ENSG00000171051: int 2357
#>   .. .. .. .. .. .. .. ..$ ENSG00000182866: int 3932
#>   .. .. .. .. .. .. .. ..$ ENSG00000131981: int 3958
#>   .. .. .. .. .. .. .. ..$ ENSG00000166002: int 56935
#>   .. .. .. .. .. .. .. ..$ ENSG00000103187: int 23406
#>   .. .. .. .. .. .. .. ..$ ENSG00000204487: int 4050
#>   .. .. .. .. .. .. .. ..$ ENSG00000271614: int 338758
#>   .. .. .. .. .. .. .. ..$ ENSG00000197629: int 219972
#>   .. .. .. .. .. .. .. ..$ ENSG00000109445: int 27309
#>   .. .. .. .. .. .. .. ..$ ENSG00000177663: int 23765
#>   .. .. .. .. .. .. .. ..$ ENSG00000132514: int 10462
#>   .. .. .. .. .. .. .. .. [list output truncated]
#>   .. .. .. .. .. .. ..$ geneBiotype           : Factor w/ 31 levels "3prime_overlapping_ncrna",..: 19 19 19 19 19 19 19 19 19 19 ...
#>   .. .. .. .. .. .. ..$ geneID                : chr [1:208] "ENSG00000204487" "ENSG00000152558" "ENSG00000121966" "ENSG00000115875" ...
#>   .. .. .. .. .. .. ..$ geneName              : Factor w/ 57063 levels "5S_rRNA","7SK",..: 22792 53320 13547 52079 28806 9704 7224 19742 15019 16133 ...
#>   .. .. .. .. .. .. ..$ seqCoordSystem        : Factor w/ 3 levels "chromosome","lrg",..: 1 1 1 1 1 1 1 1 1 1 ...
#>   .. .. .. .. .. .. ..$ gene                  : chr [1:208] "LTB" "TMEM123" "CXCR4" "SRSF7" ...
#>   .. .. .. .. .. .. ..$ gene.mean             : num [1:208] 5.24 3.92 4.54 4.63 3.36 ...
#>   .. .. .. .. .. .. ..$ gene.dispersion       : num [1:208] 6.48 5.1 5.69 7.16 5.77 ...
#>   .. .. .. .. .. .. ..$ gene.dispersion.scaled: num [1:208(1d)] 1.66 -0.553 0.185 1.619 1.125 ...
#>   .. .. .. .. .. .. .. ..- attr(*, "dimnames")=List of 1
#>   .. .. .. .. .. .. .. .. ..$ : chr [1:208] "LTB" "TMEM123" "CXCR4" "SRSF7" ...
#>   .. .. .. .. .. ..@ elementType    : chr "ANY"
#>   .. .. .. .. .. ..@ elementMetadata: NULL
#>   .. .. .. .. .. ..@ metadata       : list()
#>   .. .. .. ..@ elementType    : chr "ANY"
#>   .. .. .. ..@ metadata       :List of 8
#>   .. .. .. .. ..$ organism : chr "Homo sapiens"
#>   .. .. .. .. ..$ ensembldb:'data.frame':	12 obs. of  2 variables:
#>   .. .. .. .. .. ..$ name : chr [1:12] "Db type" "Type of Gene ID" "Supporting package" "Db created by" ...
#>   .. .. .. .. .. ..$ value: chr [1:12] "EnsDb" "Ensembl Gene ID" "ensembldb" "ensembldb package from Bioconductor" ...
#>   .. .. .. .. ..$ build    : chr "GRCh37"
#>   .. .. .. .. ..$ release  : chr "75"
#>   .. .. .. .. ..$ level    : chr "genes"
#>   .. .. .. .. ..$ version  :Classes 'package_version', 'numeric_version'  hidden list of 1
#>   .. .. .. .. .. ..$ : int [1:3] 0 99 0
#>   .. .. .. .. ..$ date     : Date[1:1], format: "2018-10-18"
#>   .. .. .. .. ..$ id       : chr "EnsDb.Hsapiens.v75"
#>   ..@ elementType    : chr "ANY"
#>   ..@ elementMetadata: NULL
#>   ..@ metadata       :List of 4
#>   .. ..$ alpha      : num 0.05
#>   .. ..$ version    :Classes 'package_version', 'numeric_version'  hidden list of 1
#>   .. .. ..$ : int [1:3] 0 2 0
#>   .. ..$ date       : Date[1:1], format: "2018-10-19"
#>   .. ..$ sessionInfo:List of 2
#>   .. .. ..$ platform:List of 9
#>   .. .. .. ..$ version : chr "R version 3.5.1 (2018-07-02)"
#>   .. .. .. ..$ os      : chr "macOS  10.14"
#>   .. .. .. ..$ system  : chr "x86_64, darwin15.6.0"
#>   .. .. .. ..$ ui      : chr "X11"
#>   .. .. .. ..$ language: chr "(EN)"
#>   .. .. .. ..$ collate : chr "en_US.UTF-8"
#>   .. .. .. ..$ ctype   : chr "en_US.UTF-8"
#>   .. .. .. ..$ tz      : chr "America/New_York"
#>   .. .. .. ..$ date    : chr "2018-10-19"
#>   .. .. .. ..- attr(*, "class")= chr "platform_info"
#>   .. .. ..$ packages:Classes 'packages_info' and 'data.frame':	221 obs. of  11 variables:
#>   .. .. .. ..$ package      : chr [1:221] "acepack" "ADGofTest" "annotate" "AnnotationDbi" ...
#>   .. .. .. ..$ ondiskversion: chr [1:221] "1.4.1" "0.3" "1.59.1" "1.43.1" ...
#>   .. .. .. ..$ loadedversion: chr [1:221] "1.4.1" "0.3" "1.59.1" "1.43.1" ...
#>   .. .. .. ..$ path         : chr [1:221] "/Users/mike/R/library/3.5-bioc-3.8-devel/acepack" "/Users/mike/R/library/3.5-bioc-3.8-devel/ADGofTest" "/Users/mike/R/library/3.5-bioc-3.8-devel/annotate" "/Users/mike/R/library/3.5-bioc-3.8-devel/AnnotationDbi" ...
#>   .. .. .. ..$ loadedpath   : chr [1:221] "/Users/mike/R/library/3.5-bioc-3.8-devel/acepack" "/Users/mike/R/library/3.5-bioc-3.8-devel/ADGofTest" "/Users/mike/R/library/3.5-bioc-3.8-devel/annotate" "/Users/mike/R/library/3.5-bioc-3.8-devel/AnnotationDbi" ...
#>   .. .. .. ..$ attached     : logi [1:221] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   .. .. .. ..$ is_base      : logi [1:221] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   .. .. .. ..$ date         : chr [1:221] "2016-10-29" "2011-12-28" "2018-07-19" "2018-06-13" ...
#>   .. .. .. ..$ source       : chr [1:221] "CRAN (R 3.5.0)" "CRAN (R 3.5.0)" "Bioconductor" "Bioconductor" ...
#>   .. .. .. ..$ md5ok        : logi [1:221] NA NA NA NA NA NA ...
#>   .. .. .. ..$ library      : Factor w/ 2 levels "/Users/mike/R/library/3.5-bioc-3.8-devel",..: 1 1 1 1 1 1 1 1 1 1 ...
#>   .. .. ..- attr(*, "class")= chr "session_info"

## `FindMarkers()` return.
invisible(capture.output(
    ident_3_markers <- Seurat::FindMarkers(
        object = object,
        ident.1 = "1",
        ident.2 = NULL
    )
))
ident_3_sanitized <- SeuratMarkers(
    markers = ident_3_markers,
    ranges = rowRanges(object)
)

#> `Seurat::FindMarkers()` return detected

str(ident_3_sanitized)

#> Formal class 'SeuratMarkers' [package "pointillism"] with 6 slots
#>   ..@ rownames       : NULL
#>   ..@ nrows          : int 196
#>   ..@ listData       :List of 7
#>   .. ..$ name    : chr [1:196] "TYMP" "LYZ" "CST3" "CD14" ...
#>   .. ..$ avgLogFC: num [1:196] 1.77 2.27 1.71 3.19 1.55 ...
#>   .. ..$ pct1    : num [1:196] 0.828 0.862 0.862 0.483 0.862 0.69 0.828 0.621 0.448 0.655 ...
#>   .. ..$ pct2    : num [1:196] 0.196 0.51 0.353 0 0.333 0.118 0.216 0.137 0.059 0.235 ...
#>   .. ..$ pvalue  : num [1:196] 1.13e-08 4.60e-08 6.73e-08 7.13e-08 1.02e-07 ...
#>   .. ..$ padj    : num [1:196] 2.59e-06 1.06e-05 1.55e-05 1.64e-05 2.35e-05 ...
#>   .. ..$ ranges  :Formal class 'GRanges' [package "GenomicRanges"] with 7 slots
#>   .. .. .. ..@ seqnames       :Formal class 'Rle' [package "S4Vectors"] with 4 slots
#>   .. .. .. .. .. ..@ values         : Factor w/ 273 levels "X","20","1","6",..: 19 7 2 18 8 10 4 1 8 3 ...
#>   .. .. .. .. .. ..@ lengths        : int [1:185] 1 1 1 1 1 1 1 1 1 1 ...
#>   .. .. .. .. .. ..@ elementMetadata: NULL
#>   .. .. .. .. .. ..@ metadata       : list()
#>   .. .. .. ..@ ranges         :Formal class 'IRanges' [package "IRanges"] with 6 slots
#>   .. .. .. .. .. ..@ start          : int [1:196] 50964181 69742121 23608534 140011313 67351066 42422614 31553901 47483612 59939081 153330330 ...
#>   .. .. .. .. .. ..@ width          : int [1:196] 4305 5894 10577 1974 3066 7857 2786 6093 13059 3174 ...
#>   .. .. .. .. .. ..@ NAMES          : chr [1:196] "TYMP" "LYZ" "CST3" "CD14" ...
#>   .. .. .. .. .. ..@ elementType    : chr "ANY"
#>   .. .. .. .. .. ..@ elementMetadata: NULL
#>   .. .. .. .. .. ..@ metadata       : list()
#>   .. .. .. ..@ strand         :Formal class 'Rle' [package "S4Vectors"] with 4 slots
#>   .. .. .. .. .. ..@ values         : Factor w/ 3 levels "+","-","*": 2 1 2 1 2 1 2 1 2 1 ...
#>   .. .. .. .. .. ..@ lengths        : int [1:103] 1 1 2 3 2 2 3 1 4 1 ...
#>   .. .. .. .. .. ..@ elementMetadata: NULL
#>   .. .. .. .. .. ..@ metadata       : list()
#>   .. .. .. ..@ seqinfo        :Formal class 'Seqinfo' [package "GenomeInfoDb"] with 4 slots
#>   .. .. .. .. .. ..@ seqnames   : chr [1:273] "X" "20" "1" "6" ...
#>   .. .. .. .. .. ..@ seqlengths : int [1:273] 155270560 63025520 249250621 171115067 198022430 159138663 133851895 135006516 191154276 81195210 ...
#>   .. .. .. .. .. ..@ is_circular: logi [1:273] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   .. .. .. .. .. ..@ genome     : chr [1:273] "GRCh37" "GRCh37" "GRCh37" "GRCh37" ...
#>   .. .. .. ..@ elementMetadata:Formal class 'DataFrame' [package "S4Vectors"] with 6 slots
#>   .. .. .. .. .. ..@ rownames       : NULL
#>   .. .. .. .. .. ..@ nrows          : int 196
#>   .. .. .. .. .. ..@ listData       :List of 10
#>   .. .. .. .. .. .. ..$ broadClass            : Factor w/ 8 levels "coding","ig",..: 1 1 1 1 1 1 1 1 1 1 ...
#>   .. .. .. .. .. .. .. ..- attr(*, "names")= chr [1:196] "ENSG00000025708" "ENSG00000090382" "ENSG00000101439" "ENSG00000170458" ...
#>   .. .. .. .. .. .. ..$ entrezID              :List of 196
#>   .. .. .. .. .. .. .. ..$ ENSG00000025708: int 1890
#>   .. .. .. .. .. .. .. ..$ ENSG00000090382: int 4069
#>   .. .. .. .. .. .. .. ..$ ENSG00000101439: int 1471
#>   .. .. .. .. .. .. .. ..$ ENSG00000170458: int 929
#>   .. .. .. .. .. .. .. ..$ ENSG00000084207: int 2950
#>   .. .. .. .. .. .. .. ..$ ENSG00000030582: int 2896
#>   .. .. .. .. .. .. .. ..$ ENSG00000204482: int 7940
#>   .. .. .. .. .. .. .. ..$ ENSG00000126759: int 5199
#>   .. .. .. .. .. .. .. ..$ ENSG00000110077: int 64231
#>   .. .. .. .. .. .. .. ..$ ENSG00000163220: int 6280
#>   .. .. .. .. .. .. .. ..$ ENSG00000215048: int 3115
#>   .. .. .. .. .. .. .. ..$ ENSG00000113269: int 55819
#>   .. .. .. .. .. .. .. ..$ ENSG00000161570: int 6352
#>   .. .. .. .. .. .. .. ..$ ENSG00000168384: int 3113
#>   .. .. .. .. .. .. .. ..$ ENSG00000204472: int 199
#>   .. .. .. .. .. .. .. ..$ ENSG00000141505: int 432
#>   .. .. .. .. .. .. .. ..$ ENSG00000197746: int 5660
#>   .. .. .. .. .. .. .. ..$ ENSG00000143546: int 6279
#>   .. .. .. .. .. .. .. ..$ ENSG00000100079: int 3957
#>   .. .. .. .. .. .. .. ..$ ENSG00000100097: int 3956
#>   .. .. .. .. .. .. .. ..$ ENSG00000015475: int 637
#>   .. .. .. .. .. .. .. ..$ ENSG00000142089: int 10410
#>   .. .. .. .. .. .. .. ..$ ENSG00000162444: int 116362
#>   .. .. .. .. .. .. .. ..$ ENSG00000104870: int 2217
#>   .. .. .. .. .. .. .. ..$ ENSG00000179639: int 2205
#>   .. .. .. .. .. .. .. ..$ ENSG00000085265: int 2219
#>   .. .. .. .. .. .. .. ..$ ENSG00000204397: int 114769
#>   .. .. .. .. .. .. .. ..$ ENSG00000204287: int 3122
#>   .. .. .. .. .. .. .. ..$ ENSG00000163191: int 6282
#>   .. .. .. .. .. .. .. ..$ ENSG00000100906: int 4792
#>   .. .. .. .. .. .. .. ..$ ENSG00000116741: int 5997
#>   .. .. .. .. .. .. .. ..$ ENSG00000196126: int [1:4] 100996809 100507714 100507709 3123
#>   .. .. .. .. .. .. .. ..$ ENSG00000158869: int 2207
#>   .. .. .. .. .. .. .. ..$ ENSG00000197249: int 5265
#>   .. .. .. .. .. .. .. ..$ ENSG00000197405: int 728
#>   .. .. .. .. .. .. .. ..$ ENSG00000216490: int 10437
#>   .. .. .. .. .. .. .. ..$ ENSG00000145220: int 55646
#>   .. .. .. .. .. .. .. ..$ ENSG00000226264: int 3109
#>   .. .. .. .. .. .. .. ..$ ENSG00000204257: int 3108
#>   .. .. .. .. .. .. .. ..$ ENSG00000164733: int 1508
#>   .. .. .. .. .. .. .. ..$ ENSG00000173762: int 924
#>   .. .. .. .. .. .. .. ..$ ENSG00000101336: int 3055
#>   .. .. .. .. .. .. .. ..$ ENSG00000198502: int 3127
#>   .. .. .. .. .. .. .. ..$ ENSG00000011600: int 7305
#>   .. .. .. .. .. .. .. ..$ ENSG00000197766: int 1675
#>   .. .. .. .. .. .. .. ..$ ENSG00000171051: int 2357
#>   .. .. .. .. .. .. .. ..$ ENSG00000182866: int 3932
#>   .. .. .. .. .. .. .. ..$ ENSG00000131981: int 3958
#>   .. .. .. .. .. .. .. ..$ ENSG00000166002: int 56935
#>   .. .. .. .. .. .. .. ..$ ENSG00000103187: int 23406
#>   .. .. .. .. .. .. .. ..$ ENSG00000204487: int 4050
#>   .. .. .. .. .. .. .. ..$ ENSG00000271614: int 338758
#>   .. .. .. .. .. .. .. ..$ ENSG00000197629: int 219972
#>   .. .. .. .. .. .. .. ..$ ENSG00000109445: int 27309
#>   .. .. .. .. .. .. .. ..$ ENSG00000177663: int 23765
#>   .. .. .. .. .. .. .. ..$ ENSG00000132514: int 10462
#>   .. .. .. .. .. .. .. ..$ ENSG00000175489: int 126364
#>   .. .. .. .. .. .. .. ..$ ENSG00000126883: int 8021
#>   .. .. .. .. .. .. .. ..$ ENSG00000130066: int 6303
#>   .. .. .. .. .. .. .. ..$ ENSG00000163131: int 1520
#>   .. .. .. .. .. .. .. ..$ ENSG00000129226: int 968
#>   .. .. .. .. .. .. .. ..$ ENSG00000100300: int 706
#>   .. .. .. .. .. .. .. ..$ ENSG00000167286: int 915
#>   .. .. .. .. .. .. .. ..$ ENSG00000185896: int 3916
#>   .. .. .. .. .. .. .. ..$ ENSG00000168961: int 3965
#>   .. .. .. .. .. .. .. ..$ ENSG00000133321: int 5920
#>   .. .. .. .. .. .. .. ..$ ENSG00000145649: int 3001
#>   .. .. .. .. .. .. .. ..$ ENSG00000130489: int 9997
#>   .. .. .. .. .. .. .. ..$ ENSG00000113140: int 6678
#>   .. .. .. .. .. .. .. ..$ ENSG00000180573: int 8334
#>   .. .. .. .. .. .. .. ..$ ENSG00000112799: int 9450
#>   .. .. .. .. .. .. .. ..$ ENSG00000177156: int 6888
#>   .. .. .. .. .. .. .. ..$ ENSG00000090104: int 5996
#>   .. .. .. .. .. .. .. ..$ ENSG00000125538: int 3553
#>   .. .. .. .. .. .. .. ..$ ENSG00000120885: int 1191
#>   .. .. .. .. .. .. .. ..$ ENSG00000127920: int 2791
#>   .. .. .. .. .. .. .. ..$ ENSG00000008517: int 9235
#>   .. .. .. .. .. .. .. ..$ ENSG00000179344: int [1:2] 101060835 3119
#>   .. .. .. .. .. .. .. ..$ ENSG00000072818: int 9744
#>   .. .. .. .. .. .. .. ..$ ENSG00000163736: int 5473
#>   .. .. .. .. .. .. .. ..$ ENSG00000166681: int 27018
#>   .. .. .. .. .. .. .. ..$ ENSG00000005961: int 3674
#>   .. .. .. .. .. .. .. ..$ ENSG00000139187: int 10219
#>   .. .. .. .. .. .. .. ..$ ENSG00000169704: int 2815
#>   .. .. .. .. .. .. .. ..$ ENSG00000101162: int 81027
#>   .. .. .. .. .. .. .. ..$ ENSG00000182718: int 302
#>   .. .. .. .. .. .. .. ..$ ENSG00000198821: int 919
#>   .. .. .. .. .. .. .. ..$ ENSG00000180644: int 5551
#>   .. .. .. .. .. .. .. ..$ ENSG00000169583: int 9022
#>   .. .. .. .. .. .. .. ..$ ENSG00000161911: int 340205
#>   .. .. .. .. .. .. .. ..$ ENSG00000100385: int 3560
#>   .. .. .. .. .. .. .. ..$ ENSG00000101856: int 10857
#>   .. .. .. .. .. .. .. ..$ ENSG00000196139: int 8644
#>   .. .. .. .. .. .. .. ..$ ENSG00000100100: int 113791
#>   .. .. .. .. .. .. .. ..$ ENSG00000136732: int 2995
#>   .. .. .. .. .. .. .. ..$ ENSG00000204276: int 3118
#>   .. .. .. .. .. .. .. ..$ ENSG00000196735: int [1:2] 100509457 3117
#>   .. .. .. .. .. .. .. ..$ ENSG00000168497: int 8436
#>   .. .. .. .. .. .. .. ..$ ENSG00000197540: int 3004
#>   .. .. .. .. .. .. .. .. [list output truncated]
#>   .. .. .. .. .. .. ..$ geneBiotype           : Factor w/ 31 levels "3prime_overlapping_ncrna",..: 19 19 19 19 19 19 19 19 19 19 ...
#>   .. .. .. .. .. .. ..$ geneID                : chr [1:196] "ENSG00000025708" "ENSG00000090382" "ENSG00000101439" "ENSG00000170458" ...
#>   .. .. .. .. .. .. ..$ geneName              : Factor w/ 57063 levels "5S_rRNA","7SK",..: 54729 22870 11204 9653 17745 17598 22788 10111 25310 49699 ...
#>   .. .. .. .. .. .. ..$ seqCoordSystem        : Factor w/ 3 levels "chromosome","lrg",..: 1 1 1 1 1 1 1 1 1 1 ...
#>   .. .. .. .. .. .. ..$ gene                  : chr [1:196] "TYMP" "LYZ" "CST3" "CD14" ...
#>   .. .. .. .. .. .. ..$ gene.mean             : num [1:196] 3.85 5.97 5.59 2.25 4.01 ...
#>   .. .. .. .. .. .. ..$ gene.dispersion       : num [1:196] 4.52 6.8 6.24 4.34 4.51 ...
#>   .. .. .. .. .. .. ..$ gene.dispersion.scaled: num [1:196(1d)] -1.46 0.707 -0.707 -0.518 -1.304 ...
#>   .. .. .. .. .. .. .. ..- attr(*, "dimnames")=List of 1
#>   .. .. .. .. .. .. .. .. ..$ : chr [1:196] "TYMP" "LYZ" "CST3" "CD14" ...
#>   .. .. .. .. .. ..@ elementType    : chr "ANY"
#>   .. .. .. .. .. ..@ elementMetadata: NULL
#>   .. .. .. .. .. ..@ metadata       : list()
#>   .. .. .. ..@ elementType    : chr "ANY"
#>   .. .. .. ..@ metadata       :List of 8
#>   .. .. .. .. ..$ organism : chr "Homo sapiens"
#>   .. .. .. .. ..$ ensembldb:'data.frame':	12 obs. of  2 variables:
#>   .. .. .. .. .. ..$ name : chr [1:12] "Db type" "Type of Gene ID" "Supporting package" "Db created by" ...
#>   .. .. .. .. .. ..$ value: chr [1:12] "EnsDb" "Ensembl Gene ID" "ensembldb" "ensembldb package from Bioconductor" ...
#>   .. .. .. .. ..$ build    : chr "GRCh37"
#>   .. .. .. .. ..$ release  : chr "75"
#>   .. .. .. .. ..$ level    : chr "genes"
#>   .. .. .. .. ..$ version  :Classes 'package_version', 'numeric_version'  hidden list of 1
#>   .. .. .. .. .. ..$ : int [1:3] 0 99 0
#>   .. .. .. .. ..$ date     : Date[1:1], format: "2018-10-18"
#>   .. .. .. .. ..$ id       : chr "EnsDb.Hsapiens.v75"
#>   ..@ elementType    : chr "ANY"
#>   ..@ elementMetadata: NULL
#>   ..@ metadata       :List of 4
#>   .. ..$ alpha      : num 0.05
#>   .. ..$ version    :Classes 'package_version', 'numeric_version'  hidden list of 1
#>   .. .. ..$ : int [1:3] 0 2 0
#>   .. ..$ date       : Date[1:1], format: "2018-10-19"
#>   .. ..$ sessionInfo:List of 2
#>   .. .. ..$ platform:List of 9
#>   .. .. .. ..$ version : chr "R version 3.5.1 (2018-07-02)"
#>   .. .. .. ..$ os      : chr "macOS  10.14"
#>   .. .. .. ..$ system  : chr "x86_64, darwin15.6.0"
#>   .. .. .. ..$ ui      : chr "X11"
#>   .. .. .. ..$ language: chr "(EN)"
#>   .. .. .. ..$ collate : chr "en_US.UTF-8"
#>   .. .. .. ..$ ctype   : chr "en_US.UTF-8"
#>   .. .. .. ..$ tz      : chr "America/New_York"
#>   .. .. .. ..$ date    : chr "2018-10-19"
#>   .. .. .. ..- attr(*, "class")= chr "platform_info"
#>   .. .. ..$ packages:Classes 'packages_info' and 'data.frame':	221 obs. of  11 variables:
#>   .. .. .. ..$ package      : chr [1:221] "acepack" "ADGofTest" "annotate" "AnnotationDbi" ...
#>   .. .. .. ..$ ondiskversion: chr [1:221] "1.4.1" "0.3" "1.59.1" "1.43.1" ...
#>   .. .. .. ..$ loadedversion: chr [1:221] "1.4.1" "0.3" "1.59.1" "1.43.1" ...
#>   .. .. .. ..$ path         : chr [1:221] "/Users/mike/R/library/3.5-bioc-3.8-devel/acepack" "/Users/mike/R/library/3.5-bioc-3.8-devel/ADGofTest" "/Users/mike/R/library/3.5-bioc-3.8-devel/annotate" "/Users/mike/R/library/3.5-bioc-3.8-devel/AnnotationDbi" ...
#>   .. .. .. ..$ loadedpath   : chr [1:221] "/Users/mike/R/library/3.5-bioc-3.8-devel/acepack" "/Users/mike/R/library/3.5-bioc-3.8-devel/ADGofTest" "/Users/mike/R/library/3.5-bioc-3.8-devel/annotate" "/Users/mike/R/library/3.5-bioc-3.8-devel/AnnotationDbi" ...
#>   .. .. .. ..$ attached     : logi [1:221] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   .. .. .. ..$ is_base      : logi [1:221] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   .. .. .. ..$ date         : chr [1:221] "2016-10-29" "2011-12-28" "2018-07-19" "2018-06-13" ...
#>   .. .. .. ..$ source       : chr [1:221] "CRAN (R 3.5.0)" "CRAN (R 3.5.0)" "Bioconductor" "Bioconductor" ...
#>   .. .. .. ..$ md5ok        : logi [1:221] NA NA NA NA NA NA ...
#>   .. .. .. ..$ library      : Factor w/ 2 levels "/Users/mike/R/library/3.5-bioc-3.8-devel",..: 1 1 1 1 1 1 1 1 1 1 ...
#>   .. .. ..- attr(*, "class")= chr "session_info"