Differential Expression

Perform pairwise differential expression across groups of cells by fitting to a zero-inflated negative binomial (ZINB) model using the zinbwave package. Currently supports edgeR and DESeq2 as DE callers.

diffExp(object, ...)

# S4 method for SingleCellExperiment
diffExp(object, numerator, denominator,
  caller = c("edgeR", "DESeq2"), minCellsPerGene = 25L,
  minCountsPerCell = 5L)

# S4 method for seurat
diffExp(object, numerator, denominator,
  caller = c("edgeR", "DESeq2"), minCellsPerGene = 25L,
  minCountsPerCell = 5L)

Arguments

object
Object.

...
Additional arguments.

numerator
character. Cells to use in the numerator of the contrast (e.g. treatment).

denominator
character. Cells to use in the denominator of the contrast (e.g. control).

caller
string. Package to use for differential expression calling. Defaults to "edgeR" (faster for large datasets) but "DESeq2" is also supported.

minCellsPerGene
scalar integer. The minimum number of cells where a gene is expressed, to pass low expression filtering. Set to 0 to disable (not recommended).

minCountsPerCell
scalar integer. Minimum number of counts per cell for a gene to pass low expression filtering. The number of cells is defined by minCellsPerGene. Set to 0 to disable (not recommended).

Value

Varies depending on the caller argument:

caller = "edgeR": DEGLRT.

caller = "DESeq2": Unshrunken DESeqResults. Use lfcShrink() if shrunken results are desired.

Details

This step will run a lot faster if you pre-calculate the ZINB weights using the runZinbwave() function, which will stash the weights into the assays() slot of the object. Running zinbwave across the entire set of filtered cells also has greater sensitivity for weight calculations.

Note

We are currently recommending the ZINB-WaVE method over zingeR, since it is faster, and has been show to be more sensitive for most single-cell RNA-seq datasets.

zinbwave

We are currently using an epsilon setting of 1e12, as recommended by the ZINB-WaVE integration paper. For more information on the zinbwave package, refer to these materials:

zinbwave paper.

zinbwave vignette.

zinbwave-DESeq2 workflow.

edgeR

After estimation of the dispersions and posterior probabilities, the glmWeightedF() function is used for statistical inference. This is an adapted function from the glmLRT() function of edgeR. It uses an F-test for which the denominator degrees of freedom are by default adjusted according to the downweighting of excess zeros (ZI = TRUE). Also, independent filtering can be performed on the obtained p-values (independentFiltering = TRUE). We use the independent filtering strategy that was originally implemented in DESeq2. By default, the average fitted values are used as a filter criterion.

DESeq2

We're providing preliminary support for DESeq2 as the differential expression caller. It is currently considerably slower for large datasets than edgeR.

We're trying to follow the conventions used in DESeq2 for contrasts, defining the name of the factor in the design formula, numerator, and denominator level for the fold change calculations. See DESeq2::results() for more information.

Seurat conventions

Note that Seurat currently uses the convention cells.1 for the numerator and cells.2 for the denominator. See Seurat::DiffExpTest() for additional information.

See also

Examples

library(bcbioSingleCell)
object <- sce_small[seq_len(100L), seq_len(100L)]
object <- filterCells(object)

#> Parameters:
#>   - >= 0 UMIs per cell
#>   - <= Inf UMIs per cell
#>   - >= 0 genes per cell
#>   - <= Inf genes per cell
#>   - >= 0 novelty score
#>   - <= 1 mitochondrial abundance
#>   - == Inf cells per sample
#>   - >= 1 cells per gene
#> ================================================================================
#> Cells:
#>      100 cells | prefilter
#>      100 cells | minUMIs >= 0
#>      100 cells | maxUMIs <= Inf
#>      100 cells | minGenes >= 0
#>      100 cells | maxGenes <= Inf
#>      100 cells | minNovelty <= 0
#>      100 cells | maxMitoRatio <= 1
#>      100 cells | nCells == Inf
#> 
#> group1 group2 
#>     49     51 
#> ================================================================================
#> Genes:
#>      100 genes | prefilter
#>      100 genes | minCellsPerGene >= 1
#> ================================================================================
#> Summary:
#>   - 100 of 100 cells passed filtering (100%)
#>   - 100 of 100 genes passed filtering (100%)

numerator <- colnames(object)[object$group == "group2"]
glimpse(numerator)

#>  chr [1:51] "cell10" "cell100" "cell11" "cell12" "cell13" "cell15" "cell16" ...

denominator <- colnames(object)[object$group == "group1"]
glimpse(denominator)

#>  chr [1:49] "cell1" "cell14" "cell17" "cell2" "cell20" "cell21" "cell22" ...

# edgeR
x <- suppressMessages(diffExp(
    object = object,
    numerator = numerator,
    denominator = denominator,
    caller = "edgeR"
))
class(x)

#> [1] "DGELRT"
#> attr(,"package")
#> [1] "edgeR"

summary(x)

#>                       Length Class            Mode     
#> coefficients           160   -none-           numeric  
#> fitted.values         8000   -none-           numeric  
#> deviance                80   -none-           numeric  
#> method                   1   -none-           character
#> unshrunk.coefficients  160   -none-           numeric  
#> df.residual             80   -none-           numeric  
#> design                 200   -none-           numeric  
#> offset                 100   CompressedMatrix numeric  
#> dispersion              80   -none-           numeric  
#> weights               8000   -none-           numeric  
#> prior.count              1   -none-           numeric  
#> samples                  3   data.frame       list     
#> prior.df                 1   -none-           numeric  
#> AveLogCPM               80   -none-           numeric  
#> df.total                80   -none-           numeric  
#> table                    4   data.frame       list     
#> comparison               1   -none-           character
#> df.test                 80   -none-           numeric  

# DESeq2
x <- suppressMessages(diffExp(
    object = object,
    numerator = numerator,
    denominator = denominator,
    caller = "DESeq2"
))
class(x)

#> [1] "DESeqResults"
#> attr(,"package")
#> [1] "DESeq2"

summary(x)

#>       Length        Class         Mode 
#>            6 DESeqResults           S4