Fetch Data Functions

Fetch Data Functions

fetchGeneData(object, ...)

fetchReducedDimData(object, ...)

fetchReducedDimExpressionData(object, ...)

# S4 method for SingleCellExperiment
fetchGeneData(object, genes,
  gene2symbol = FALSE)

# S4 method for SingleCellExperiment
fetchReducedDimData(object, reducedDim,
  dimsUse = c(1L, 2L), interestingGroups)

# S4 method for SingleCellExperiment
fetchReducedDimExpressionData(object,
  genes, reducedDim)

Arguments

object
Object.

...
Additional arguments.

genes
character. Gene identifiers. Must match the rownames of the object.

gene2symbol
data.frame. Gene-to-symbol mappings. Columns must contain geneID and geneName.

reducedDim
string. Name of reduced dimension matrix slotted in reducedDims(). Includes TNSE, UMAP, PCA, for example.

dimsUse
integer. Vector of length 2 that denotes the columns from the reduced dimension matrix to use for centerX and centerY column calculations. Defaults the first and second dimensions.

interestingGroups
character or NULL. Character vector of interesting groups. Must be formatted in camel case and intersect with sampleData() colnames.

Value

fetchGeneData(): matrix.

fetchTSNEExpressionData(): data.frame containing t-SNE coordinates, sample metadata, and aggregate marker expression values (mean, median, and sum).

Other functions: data.frame, containing metrics.

Note

Some of these functions require "ident" to be defined in colData(). For t-SNE, UMAP, and PCA, the reduced dimensions must be defined in the reducedDims() slot.

Examples

object <- indrops_small
genes <- head(rownames(object))
glimpse(genes)

#>  chr [1:6] "ENSG00000000457" "ENSG00000002586" "ENSG00000004939" ...

# Genes
x <- fetchGeneData(object, genes = genes)
glimpse(x)

#>  num [1:500, 1:6] 1 1 4 3 3 2 5 1 3 3 ...
#>  - attr(*, "dimnames")=List of 2
#>   ..$ : chr [1:500] "multiplexed_AAAAAAAA_AAACACTA_ATCCCACG" "multiplexed_AAAAAAAA_AAACACTA_CTTAGGTA" "multiplexed_AAAAAAAA_AAACACTA_CTTCGATT" "multiplexed_AAAAAAAA_AAACCGCC_GAGTTTCG" ...
#>   ..$ : chr [1:6] "ENSG00000000457" "ENSG00000002586" "ENSG00000004939" "ENSG00000005022" ...

# Reduced dim coordinates
x <- fetchReducedDimExpressionData(
    object = object,
    genes = genes,
    reducedDim = "TSNE"
)

#> Error in .assertHasIdent(object): is_subset : The element 'ident' in "ident" is not in colnames(colData(object)).

glimpse(x)

#>  num [1:500, 1:6] 1 1 4 3 3 2 5 1 3 3 ...
#>  - attr(*, "dimnames")=List of 2
#>   ..$ : chr [1:500] "multiplexed_AAAAAAAA_AAACACTA_ATCCCACG" "multiplexed_AAAAAAAA_AAACACTA_CTTAGGTA" "multiplexed_AAAAAAAA_AAACACTA_CTTCGATT" "multiplexed_AAAAAAAA_AAACCGCC_GAGTTTCG" ...
#>   ..$ : chr [1:6] "ENSG00000000457" "ENSG00000002586" "ENSG00000004939" "ENSG00000005022" ...