Find cluster-specific marker genes

Find cluster-specific marker genes

findMarkers(object, ...)

# S4 method for SingleCellExperiment
findMarkers(object, clusters = NULL, ...)

# S4 method for Seurat
findMarkers(object, clusters = NULL, ...)

Arguments

object
Object.

clusters
character or NULL. Cluster identifiers. Must correspond to values in clusters(). Note that Seurat uses zero-indexed IDs by default (e.g. 0, 1, 2, ...). If left NULL (default), all clusters will be analyzed. Looping across clusters manually here can avoid memory issues on laptops and other machines with limited amounts of RAM.

...
Passthrough arguments to diffExp().

Value

list containing:

caller = "edgeR": DGELRT.

caller = "DESeq2": DESeqResults.

Note

Updated 2020-01-30.

Examples

data(Seurat, package = "AcidTest")

## Seurat ====
object <- Seurat
x <- findMarkers(object, caller = "edgeR")

#> 

#> ── `findMarkers()` ─────────────────────────────────────────────────────────────

#> ℹ 3 clusters detected.

#> 

#> ── Cluster 0 ──

#> 

#> → Performing differential expression with edgeR.

#>     ● Total: 80 cells

#>     ● Numerator: 36 cells

#>     ● Denominator: 44 cells

#> → Applying gene expression low pass filter.

#>     ℹ Requiring at least 1 cell with counts of 1 or more per gene.

#>     ℹ 230 of 230 genes passed filter.

#> → Applying cell low pass filter.

#> → Running edgeR.

#> 

#> ── Cluster 1 ──

#> 

#> → Performing differential expression with edgeR.

#>     ● Total: 80 cells

#>     ● Numerator: 25 cells

#>     ● Denominator: 55 cells

#> → Applying gene expression low pass filter.

#>     ℹ Requiring at least 1 cell with counts of 1 or more per gene.

#>     ℹ 230 of 230 genes passed filter.

#> → Applying cell low pass filter.

#> → Running edgeR.

#> 

#> ── Cluster 2 ──

#> 

#> → Performing differential expression with edgeR.

#>     ● Total: 80 cells

#>     ● Numerator: 19 cells

#>     ● Denominator: 61 cells

#> → Applying gene expression low pass filter.

#>     ℹ Requiring at least 1 cell with counts of 1 or more per gene.

#>     ℹ 230 of 230 genes passed filter.

#> → Applying cell low pass filter.

#> → Running edgeR.

class(x)

#> [1] "list"

lapply(x, class)

#> $`0`
#> [1] "DGELRT"
#> attr(,"package")
#> [1] "edgeR"
#> 
#> $`1`
#> [1] "DGELRT"
#> attr(,"package")
#> [1] "edgeR"
#> 
#> $`2`
#> [1] "DGELRT"
#> attr(,"package")
#> [1] "edgeR"
#>