Known Markers

Class containing known markers detected.

KnownMarkers(markers, known, ...)

# S4 method for SeuratMarkersPerCluster,CellTypeMarkers
KnownMarkers(markers,
  known, promiscuousThreshold = 0L)

Arguments

markers
SeuratMarkers or SeuratMarkersPerCluster.

known
CellTypeMarkers. Grouped by cellType column. Known markers data.frame imported by readCellTypeMarkers() or pulled from internal cell cycle markers data.

...
Additional arguments.

promiscuousThreshold
scalar integer. Minimum number of clusters required to consider a gene marker promiscuous. Set to 0 to disable promiscuous marker filtering.

Value

KnownMarkers.

Note

Both the markers and known objects must contain Ensembl gene identifiers in the geneID column. We must avoid any matching operations based on the gene names, since these change often and can mismatch easily.

See also

Examples

data(all_markers_small, cell_type_markers)
x <- KnownMarkers(
    markers = all_markers_small,
    known = cell_type_markers$homoSapiens
)

#> Grouping by cluster.

#> Grouping by cellType.

summary(x)

#>       Length        Class         Mode 
#>           10 KnownMarkers           S4