Quality control metrics

Quality control metrics

metrics(object, ...)

# S4 method for cell_data_set
metrics(object, return = c("tbl_df",
  "DataFrame"))

Arguments

object
Object.

return
character(1). Return type. Uses match.arg() internally and defaults to the first argument in the character vector.

...
Additional arguments.

Value

"tbl_df": grouped_df. Tibble grouped by sampleID column.

"DataFrame": DataFrame. Row names are identical to the column names of the object, like colData().

Details

metrics() takes data stored in colData() and consistently returns a tibble grouped by sample by default (sampleID). It always returns sampleName and interestingGroups columns, even when these columns are not defined in colData. This is designed to integrate with plotting functions that use ggplot2 internally.

Note

Updated 2019-08-02.

Methods (by class)

SummarizedExperiment: Metrics are sample level. sampleID column corresponds to colnames.

SingleCellExperiment: Metrics are cell level. cellID column corresponds to colnames. Tibble is returned grouped by sample (sampleID column).

Examples

data(
    cell_data_set,
    package = "acidtest"
)

## cell_data_set ====
object <- cell_data_set
x <- metrics(object)
print(x)

#> # A tibble: 200 x 5
#> # Groups:   sampleID [1]
#>    cellID                  sampleID sampleName interestingGroups ident
#>    <chr>                   <fct>    <fct>      <fct>             <fct>
#>  1 cele_009_002_AATATTCGAA unknown  unknown    unknown           1    
#>  2 cele_005_091_TCTAGAACTA unknown  unknown    unknown           2    
#>  3 cele_007_083_ATCTGCCAAT unknown  unknown    unknown           1    
#>  4 cele_010_091_TGAACGGTAA unknown  unknown    unknown           1    
#>  5 cele_008_092_TGACCGAATC unknown  unknown    unknown           2    
#>  6 cele_005_089_AATACCAGTT unknown  unknown    unknown           2    
#>  7 cele_007_044_TGACCGAATC unknown  unknown    unknown           2    
#>  8 cele_010_091_TATCCTCAAT unknown  unknown    unknown           2    
#>  9 cele_007_082_TTATTAGTAG unknown  unknown    unknown           1    
#> 10 cele_010_060_ATTCAGGTAC unknown  unknown    unknown           2    
#> # … with 190 more rows

x <- metricsPerSample(object, fun = "mean")

#> Calculating mean per sample.

print(x)

#> # A tibble: 1 x 3
#> # Groups:   sampleID [1]
#>   sampleID sampleName interestingGroups
#>   <fct>    <fct>      <fct>            
#> 1 unknown  unknown    unknown