Plot Cell-Type-Specific Gene Markers

Plot gene expression per cell in multiple formats:

plotMarkerTSNE(): t-SNE gene expression plot.

plotDot(): Dot plot.

plotViolin(): Violin plot.

plotKnownMarkersDetected(object, ...)

plotMarker(object, ...)

plotTopMarkers(object, ...)

# S4 method for SingleCellExperiment
plotMarker(object, genes,
  reducedDim = c("TSNE", "UMAP"), expression = c("mean", "median",
  "sum"), color = getOption("bcbio.discrete.color", NULL),
  pointSize = getOption("bcbio.pointSize", 0.75),
  pointAlpha = getOption("bcbio.pointAlpha", 0.8),
  pointsAsNumbers = FALSE, label = getOption("bcbio.label", TRUE),
  labelSize = getOption("bcbio.labelSize", 6L),
  dark = getOption("bcbio.dark", FALSE), grid = getOption("bcbio.grid",
  FALSE), legend = getOption("bcbio.legend", TRUE),
  aspectRatio = getOption("bcbio.aspectRatio", 1L), title = TRUE)

# S4 method for SingleCellExperiment
plotTopMarkers(object, markers, n = 10L,
  direction = c("positive", "negative", "both"), coding = FALSE,
  reducedDim = c("TSNE", "UMAP"), headerLevel = 2L, ...)

# S4 method for SingleCellExperiment
plotKnownMarkersDetected(object, markers,
  reducedDim = c("TSNE", "UMAP"), headerLevel = 2L, ...)

Arguments

object
Object.

...
Additional arguments.

genes
character. Gene identifiers. Must match the rownames of the object.

reducedDim
string. Name of reduced dimension matrix slotted in reducedDims(). Includes TNSE, UMAP, PCA, for example.

expression
string. Calculation to apply. Uses match.arg() and defaults to the first argument in the character vector.

color
ggproto/ScaleDiscrete or NULL. Desired ggplot2 color scale. Must supply discrete values. When set to NULL, the default ggplot2 color palette will be used. If manual color definitions are desired, we recommend using ggplot2::scale_color_manual(). To set the discrete color palette globally, use options(bcbio.discrete.color = scale_color_viridis_d()).

pointSize
scalar numeric. Cell point size.

pointAlpha
scalar numeric (0-1). Alpha transparency level. Useful when there many cells in the dataset, and some cells can be masked.

pointsAsNumbers
boolean. Plot the points as numbers (TRUE) or dots (FALSE).

label
boolean. Overlay a cluster identitiy label on the plot.

labelSize
scalar integer. Size of the text label.

dark
boolean. Plot against a dark background using basejump::theme_midnight().

grid
boolean. Show major grid lines but hide axis lines.

legend
boolean. Include plot legend.

aspectRatio
scalar integer. Aspect ratio.

title
string or NULL. Plot title.

markers
grouped_df. Marker genes data.

plotTopMarkers(): must be grouped by "cluster".

plotKnownMarkersDetected(): must be grouped by "cellType".

headerLevel
scalar integer (1-7). R Markdown header level.

Value

Show graphical output. Invisibly return ggplot list.

Plot top markers

The number of markers to plot is determined by the output of the topMarkers() function. If you want to reduce the number of genes to plot, simply reassign first using that function. If necessary, we can add support for the number of genes to plot here in a future update.

See also

Examples

# MHC class II genes
title <- "MHC class II"
genes <- rownames(sce_small)[which(grepl(
    pattern = "^major histocompatibility complex, class II",
    x = rowData(sce_small)$description
))]
print(genes)

#> character(0)

plotMarker(sce_small, genes = genes)

#> Error in assertFormalInterestingGroups(object, value): The interesting groups "sampleName" are not defined as columns in `sampleData()`

plotMarker(
    object = sce_small,
    genes = genes,
    expression = "sum",
    pointsAsNumbers = TRUE,
    dark = TRUE,
    label = FALSE,
    title = title
)

#> Error in assertFormalInterestingGroups(object, value): The interesting groups "sampleName" are not defined as columns in `sampleData()`

# Top markers
markers <- topMarkers(all_markers_small, n = 1)
glimpse(markers)

#> Observations: 2
#> Variables: 18
#> $ cluster        <fct> 0, 1
#> $ avgLogFC       <dbl> 1.284217, 1.087316
#> $ pct1           <dbl> 0.962, 0.979
#> $ pct2           <dbl> 0.938, 0.865
#> $ rowname        <chr> "SGCE", "IRAK3"
#> $ pvalue         <dbl> 1.642010e-15, 1.291082e-16
#> $ padj           <dbl> 1.642010e-11, 1.291082e-12
#> $ seqnames       <fct> 7, 12
#> $ start          <int> 94524204, 66188879
#> $ end            <int> 94656572, 66254622
#> $ width          <int> 132369, 65744
#> $ strand         <fct> -, +
#> $ broadClass     <fct> coding, coding
#> $ description    <fct> sarcoglycan epsilon [Source:HGNC Symbol;Acc:HGNC:108...
#> $ geneBiotype    <fct> protein_coding, protein_coding
#> $ geneID         <chr> "ENSG00000127990", "ENSG00000090376"
#> $ geneName       <fct> SGCE, IRAK3
#> $ seqCoordSystem <fct> chromosome, chromosome

plotTopMarkers(seurat_small, markers = tail(markers, 1))

#> 
#> 
#> ## Cluster 1 {.tabset}
#> 
#> 
#> 
#> ### IRAK3
#> 

#> Error in assertFormalInterestingGroups(object, value): The interesting groups "sampleName" are not defined as columns in `sampleData()`

# Known markers detected
markers <- head(known_markers_small, n = 1)
glimpse(markers)

#> Observations: 1
#> Variables: 19
#> $ cellType       <fct> Gut Epithelial Cell
#> $ cluster        <fct> 0
#> $ geneID         <chr> "ENSG00000066405"
#> $ geneName       <chr> "CLDN18"
#> $ avgLogFC       <dbl> -0.9446141
#> $ pct1           <dbl> 0.442
#> $ pct2           <dbl> 0.688
#> $ rowname        <chr> "CLDN18"
#> $ pvalue         <dbl> 3.27838e-06
#> $ padj           <dbl> 0.0327838
#> $ seqnames       <chr> "3"
#> $ start          <int> 137998735
#> $ end            <int> 138033655
#> $ width          <int> 34921
#> $ strand         <chr> "+"
#> $ broadClass     <chr> "coding"
#> $ description    <chr> "claudin 18 [Source:HGNC Symbol;Acc:HGNC:2039]"
#> $ geneBiotype    <chr> "protein_coding"
#> $ seqCoordSystem <chr> "chromosome"

plotKnownMarkersDetected(seurat_small, markers = head(markers, 1))

#> 
#> 
#> ## Gut Epithelial Cell {.tabset}
#> 
#> 
#> 
#> ### CLDN18
#> 

#> Error in assertFormalInterestingGroups(object, value): The interesting groups "sampleName" are not defined as columns in `sampleData()`