Run Seurat

Run Seurat

runSeurat(object, ...)

# S4 method for Seurat
runSeurat(
  object,
  regressCellCycle = c("s-g2m-diff", "yes", "no"),
  varsToRegress = c("nCount_RNA", "mitoRatio"),
  dims = "auto",
  resolution = seq(from = 0.2, to = 1.2, by = 0.2),
  tsneMethod = "Rtsne",
  umapMethod = "umap-learn",
  virtualenv = "r-reticulate",
  workers = "auto"
)

# S4 method for SingleCellExperiment
runSeurat(object, ...)

Arguments

object
Object.

...
Additional arguments.

regressCellCycle
character(1).

"s-g2m-diff": Calculate the difference between S and G2/M phases and use that to regress. See CC.Difference metric in Seurat vignette.

"yes": Regress out any effects of both S and G2/M phase variable. Refer to "S.Score" and "G2M.Score" metrics in Seurat vignette.

"no": Don't calculate cell-cycle scoring and don't regress.

Refer to the Seurat cell-cycle regression vignette for details.

varsToRegress
character or NULL. Unwanted sources of variance to regress. Note that when regressCellCycle is not "no", then the corresponding cell-cycle variables are added automatically. Passes to Seurat::ScaleData internally.

dims
"auto" or integer. Dimensions of reduction to use as input for shared nearest neighbor (SNN) graph construction. When set to "auto" (default), the elbow point is calculated internally. See plotPCElbow() for details. Passes to Seurat::FindNeighbors() and Seurat::RunUMAP() internally.

resolution
numeric. Resolutions to calculate for clustering. Passes to Seurat::FindClusters() internally.

tsneMethod
character(1). t-SNE algorithm. Using the current default in Seurat.

umapMethod
character(1). UMAP algorithm. Use either "umap-learn", which requires reticulate, or "uwot". Note that Seurat 3 changed the default from umap-learn to uwot, but I prefer to call the python package directly.

virtualenv
character(1). Python virtual environment name. Only evaluated when umapMethod = "umap-learn". See reticulate section for details.

workers
"auto", integer(1), or NULL. Disable parallelization with future by setting to NULL.

Note

Updated 2020-02-21.

reticulate

A Python virtual environment (e.g. "r-reticulate") with umap-learn installed is required when umapMethod is set to "umap-learn" (default).

The reticulate package supports conda or virtual environments managed by venv/virtualenv. I've had better luck with using using virtual environments rather than attempting to use conda with this package.

Check Sys.getenv("WORKON_HOME") path, which is the current approach used in Rprofile.site.

See also