Run ZINB-WaVE

Given an object with the data, it performs dimensionality reduction using a ZINB regression model with gene and cell-level covariates.

runZinbwave(Y, ...)

# S4 method for SingleCellExperiment
runZinbwave(
  Y,
  K = 0L,
  epsilon = 1e+12,
  BPPARAM = BiocParallel::bpparam(),
  recalculate = FALSE,
  verbose = FALSE
)

Arguments

Y
The data (genes in rows, samples in columns). Currently implemented only for SummarizedExperiment.

...
Additional parameters to describe the model, see zinbModel.

K
integer. Number of latent factors. Specify K = 0 if only computing observational weights.

epsilon
nonnegative scalar. Regularization parameter.

BPPARAM
object of class bpparamClass that specifies the back-end to be used for computations. See bpparam for details.

recalculate
logical. Force recalculation of weights.

verbose
logical. Run zinbwave in verbose mode (for debugging).

Value

Modified object. Weights added to assays.

Details

zinbwave will calculate normalizedValues and weights matrices, which will be defined in the assays slot. If these values have already been calculated, the object will be returned unmodified, unless recalculate = TRUE.

Note

Updated 2020-02-21.

Parallelization

zinbwave() defaults to using BiocParallel::bpparam to register the number of cores to use (bpparam argument), but we've found that this works inconsistently across installations. Currently, we recommend using BiocParallel::SerialParam by default, which will run in serial, using a single core. On Linux or macOS, BiocParallel::MulticoreParam should work to run with multiple cores.

See also

Examples

data(SingleCellExperiment, package = "acidtest")
Y <- SingleCellExperiment
Y <- nonzeroRowsAndCols(Y)

#> ℹ Filtered zero count rows and columns:
#> - 485 / 500 rows (97%)
#> - 100 / 100 columns (100%)

Y <- runZinbwave(Y)

#> → Running zinbwave.

#> ℹ BiocParallel param registered.
#> class: MulticoreParam
#>   bpisup: FALSE; bpnworkers: 3; bptasks: 0; bpjobname: BPJOB
#>   bplog: FALSE; bpthreshold: INFO; bpstopOnError: TRUE
#>   bpRNGseed: ; bptimeout: 2592000; bpprogressbar: TRUE
#>   bpexportglobals: TRUE
#>   bplogdir: NA
#>   bpresultdir: NA
#>   cluster type: FORK

#> → `zinbFit()`: Fitting a ZINB regression model.

#> CPU time used:
#>    user  system elapsed 
#>  10.001   5.537   6.200 

#> Object of class ZinbModel. 100 samples; 485 genes. 1 sample-level
#> covariate(s) (mu); 1 sample-level covariate(s) (pi); 1 gene-level
#> covariate(s) (mu); 1 gene-level covariate(s) (pi); 0 latent
#> factor(s).

#> → `zinbwave()`: Performing dimensionality reduction using a ZINB regression model with gene and cell-level covariates.

#> CPU time used: user system elapsed 0.005 0.000 0.006