Force an object to belong to a class

Coerce an object to a given class.

CellCycleMarkers to tbl_df

S4 coercion support for creating a tbl_df from CellCycleMarkers.

CellTypeMarkers to tbl_df

S4 coercion support for creating a tbl_df from CellTypeMarkers.

SingleCellExperiment to seurat

Interally Seurat::CreateSeuratObject is called without applying any additional filtering cutoffs, since we have already defined them during our quality control analysis. Here we are passing the raw gene-level counts of the filtered cells into a new seurat class object. Use convertGenesToSymbols to convert gene IDs to names (symbols).

seurat to SingleCellExperiment

S4 coercion support for creating a SingleCellExperiment from a seurat class object. The Seurat FAQ page explains the seurat S4 class structure in detail. Internally, this method improves the basic Seurat::as.SingleCellExperiment S3 coercion method, including the object@scale.data matrix, and will keep track of stashed rowRanges and metadata if the seurat object was originally created from a SingleCellExperiment (i.e. from the bcbioSingleCell package).

seurat to RangedSummarizedExperiment

S4 coercion support for creating a RangedSummarizedExperiment from a seurat class object.

seurat to SummarizedExperiment

S4 coercion support for creating a SummarizedExperiment from a seurat class object.

SeuratMarkers to tbl_df

S4 coercion support for creating a tbl_df from a Markers object.

SeuratMarkersPerCluster to tbl_df

S4 coercion support for creating a tbl_df from SeuratMarkersPerCluster.

Description

as(object) returns the version of this object coerced to be the given Class. When used in the replacement form on the left of an assignment, the portion of the object corresponding to Class is replaced by value.

The operation of as() in either form depends on the definition of coerce methods. Methods are defined automatically when the two classes are related by inheritance; that is, when one of the classes is a subclass of the other.

Coerce methods are also predefined for basic classes (including all the types of vectors, functions and a few others).

Beyond these two sources of methods, further methods are defined by calls to the setAs function. See that documentation also for details of how coerce methods work. Use showMethods(coerce) for a list of all currently defined methods, as in the example below.

Basic Coercion Methods

Methods are pre-defined for coercing any object to one of the basic datatypes. For example, as(x, "numeric") uses the existing as.numeric function. These and all other existing methods can be listed as shown in the example.

References

Chambers, John M. (2016) Extending R, Chapman & Hall. (Chapters 9 and 10.)

See also

Examples

data(sce, package = "basejump")
data(seurat_small)

## SingleCellExperiment to seurat ====
object <- sce
print(object)

#> class: SingleCellExperiment 
#> dim: 500 100 
#> metadata(1): date
#> assays(1): counts
#> rownames(500): gene1 gene2 ... gene499 gene500
#> rowData names(5): geneID geneName geneBiotype broadClass entrezID
#> colnames(100): cell001 cell002 ... cell099 cell100
#> colData names(2): expLibSize sampleID
#> reducedDimNames(0):
#> spikeNames(0):

x <- as(object, "seurat")
class(x)

#> [1] "seurat"
#> attr(,"package")
#> [1] "Seurat"

print(x)

#> An object of class seurat in project pointillism 
#>  500 genes across 100 samples.

## seurat to SingleCellExperiment ====
x <- as(seurat_small, "SingleCellExperiment")
print(x)

#> class: SingleCellExperiment 
#> dim: 230 80 
#> metadata(2): weights varGenes
#> assays(4): counts logcounts scaleData weights
#> rownames(230): MS4A1 CD79B ... SPON2 S100B
#> rowData names(10): broadClass entrezID ... gene.dispersion
#>   gene.dispersion.scaled
#> colnames(80): ATGCCAGAACGACT CATGGCCTGTGCAT ... GGAACACTTCAGAC
#>   CTTGATTGATCTTC
#> colData names(6): nGene nUMI ... res.1 ident
#> reducedDimNames(3): PCA TSNE UMAP
#> spikeNames(0):