Differential Expression

Perform pairwise differential expression across groups of cells by fitting to a zero-inflated negative binomial (ZINB) model using the zinbwave package. Currently supports edgeR and DESeq2 as DE callers.

diffExp(object, ...)

# S4 method for SingleCellExperiment
diffExp(object, numerator, denominator,
  caller = c("edgeR", "DESeq2"), minCells = 2L, minCellsPerGene = 1L,
  minCountsPerCell = 1L, bpparam = BiocParallel::SerialParam())

# S4 method for seurat
diffExp(object, numerator, denominator,
  caller = c("edgeR", "DESeq2"), minCells = 2L, minCellsPerGene = 1L,
  minCountsPerCell = 1L, bpparam = BiocParallel::SerialParam())

Arguments

object
Object.

...
Additional arguments.

numerator
character. Cells to use in the numerator of the contrast (e.g. treatment).

denominator
character. Cells to use in the denominator of the contrast (e.g. control).

caller
string. Package to use for differential expression calling. Defaults to "edgeR" (faster for large datasets) but "DESeq2" is also supported.

minCells
scalar integer. Minimum number of cells required to perform the differential expression analysis.

minCellsPerGene
scalar integer. The minimum number of cells where a gene is expressed, to pass low expression filtering.

minCountsPerCell
scalar integer. Minimum number of counts per cell for a gene to pass low expression filtering. The number of cells is defined by minCellsPerGene.

bpparam
bpparamClass. BiocParallel parameter specifying the back-end to be used for computations. See BiocParallel::bpparam() for details. We recommend one of the following:

BiocParallel::bpparam().

BiocParallel::SerialParam().

BiocParallel::MulticoreParam(). BiocParallel: https://doi.org/doi:10.18129/B9.bioc.BiocParallel

Value

Varies depending on the caller argument:

caller = "edgeR": DEGLRT.

caller = "DESeq2": Unshrunken DESeqResults. Use lfcShrink() if shrunken results are desired.

Note

We are currently recommending the ZINB-WaVE method over zingeR, since it is faster, and has been show to be more sensitive for most single-cell RNA-seq datasets.

zinbwave

This function will run a lot faster if you pre-calculate the ZINB weights using the runZinbwave() function, which will stash the weights into the SingleCellExperiment::weights() slot of the object. Running zinbwave across the entire set of filtered cells also has greater sensitivity for weight calculations.

We are currently using an epsilon setting of 1e12, as recommended by the ZINB-WaVE integration paper. For more information on the zinbwave package, refer to these materials:

zinbwave paper.

zinbwave vignette.

zinbwave-DESeq2 workflow.

edgeR

After estimation of the dispersions and posterior probabilities, the zinbwave::glmWeightedF() function is used for statistical inference. This is an adaptation of edgeR::glmLRT(). It uses an F-test for which the denominator degrees of freedom are by default adjusted according to the downweighting of excess zeros (ZI = TRUE). Also, independent filtering can be performed on the obtained p-values (independentFiltering = TRUE). We use the independent filtering strategy that was originally implemented in DESeq2. By default, the average fitted values are used as a filter criterion.

DESeq2

We're providing preliminary support for DESeq2 as the differential expression caller. It is currently considerably slower for large datasets than edgeR.

We're trying to follow the conventions used in DESeq2 for contrasts, defining the name of the factor in the design formula, numerator, and denominator level for the fold change calculations. See DESeq2::results() for more information.

Seurat conventions

Note that Seurat currently uses the convention cells.1 for the numerator and cells.2 for the denominator. See Seurat::DiffExpTest() for additional information.

See also

Examples

data(seurat_small)
object <- seurat_small

## Calculate ZINB weights (already stashed).
## seurat_small <- runZinbwave(object)

## Compare expression in cluster 3 relative to 2.
ident <- clusterID(object)
numerator <- names(ident)[ident == "3"]
summary(numerator)

#>    Length     Class      Mode 
#>        10 character character 

denominator <- names(ident)[ident == "2"]
summary(denominator)

#>    Length     Class      Mode 
#>        12 character character 

## edgeR.
x <- diffExp(
    object = object,
    numerator = numerator,
    denominator = denominator,
    caller = "edgeR"
)

#> Performing differential expression with zinbwave-edgeR

#> Total: 22 cells
#> Numerator: 10 cells
#> Denominator: 12 cells

#> Applying gene expression low pass filter.

#> Requiring at least 1 cells with counts of 1 or more per gene

#> 187 of 230 genes passed filter

#> Running edgeR.

#>    user  system elapsed 
#>   0.160   0.005   0.168 

class(x)

#> [1] "DGELRT"
#> attr(,"package")
#> [1] "edgeR"

summary(x)

#>                       Length Class            Mode     
#> coefficients           374   -none-           numeric  
#> fitted.values         4114   -none-           numeric  
#> deviance               187   -none-           numeric  
#> method                   1   -none-           character
#> unshrunk.coefficients  374   -none-           numeric  
#> df.residual            187   -none-           numeric  
#> design                  44   -none-           numeric  
#> offset                  22   CompressedMatrix numeric  
#> dispersion             187   -none-           numeric  
#> weights               4114   -none-           numeric  
#> prior.count              1   -none-           numeric  
#> samples                  3   data.frame       list     
#> prior.df                 1   -none-           numeric  
#> AveLogCPM              187   -none-           numeric  
#> df.total               187   -none-           numeric  
#> table                    4   data.frame       list     
#> comparison               1   -none-           character
#> df.test                187   -none-           numeric  

## DESeq2.
x <- diffExp(
    object = object,
    numerator = numerator,
    denominator = denominator,
    caller = "DESeq2"
)

#> Performing differential expression with zinbwave-DESeq2

#> Total: 22 cells
#> Numerator: 10 cells
#> Denominator: 12 cells

#> Applying gene expression low pass filter.

#> Requiring at least 1 cells with counts of 1 or more per gene

#> 187 of 230 genes passed filter

#> Running DESeq2.

#> converting counts to integer mode

#> estimating size factors

#> estimating dispersions

#> gene-wise dispersion estimates

#> Warning: for 33 row(s), the weights as supplied won't allow parameter estimation, producing a
#>   degenerate design matrix. These rows have been flagged in mcols(dds)$weightsFail
#>   and treated as if the row contained all zeros (mcols(dds)$allZero set to TRUE).

#> mean-dispersion relationship

#> final dispersion estimates

#> fitting model and testing

#>    user  system elapsed 
#>   1.322   0.022   1.369 

class(x)

#> [1] "DESeqResults"
#> attr(,"package")
#> [1] "DESeq2"

summary(x)

#>       Length        Class         Mode 
#>            6 DESeqResults           S4