Perform dimensionality reduction using a ZINB regression model with gene and cell-level covariates.

Given an object with the data, it performs dimensionality reduction using a ZINB regression model with gene and cell-level covariates.

runZinbwave(Y, ...)

# S4 method for SingleCellExperiment
runZinbwave(Y, K = 0L,
  epsilon = 1e+12, BPPARAM = BiocParallel::SerialParam(),
  recalculate = FALSE, verbose = FALSE)

# S4 method for seurat
runZinbwave(Y, ...)

Arguments

Y
The data (genes in rows, samples in columns). Currently implemented only for SummarizedExperiment.

...
Additional parameters to describe the model, see zinbModel.

K
integer. Number of latent factors. Specify K = 0 if only computing observational weights.

epsilon
nonnegative scalar. Regularization parameter.

BPPARAM
object of class bpparamClass that specifies the back-end to be used for computations. See bpparam for details.

recalculate
logical. Force recalculation of weights.

verbose
logical. Run zinbwave in verbose mode (for debugging).

Value

Modified object (S4 class must extend SingleCellExperiment), with weights added to assays.

Details

zinbwave will calculate normalizedValues and weights matrices, which will be defined in the assays slot. If these values have already been calculated, the object will be returned unmodified, unless recalculate = TRUE.

Methods (by class)

seurat: Coerces to SingleCellExperiment and stashes weights in the weights slot.

Parallelization

zinbwave defaults to using BiocParallel::bpparam to register the number of cores to use (bpparam argument), but we've found that this works inconsistently across installations. Currently, we recommend using BiocParallel::SerialParam by default, which will run in serial, using a single core. On Linux or macOS, BiocParallel::MulticoreParam should work to run with multiple cores.

See also

Examples

data(seurat_small)
Y <- seurat_small
Y <- runZinbwave(Y = Y)

#> Running zinbwave.

#> Object already contains pre-calculated weights.

class(weights(Y))

#> [1] "matrix"