Calculate ZINB-WaVE Weights

zinbwave will calculate normalizedValues and weights matrices, which will be defined in the assays() slot. If these values have already been calculated, the object will be returned unmodified, unless recalculate = TRUE.

runZinbwave(Y, ...)

# S4 method for SingleCellExperiment
runZinbwave(Y, K = 0L,
  epsilon = 1e+12, BPPARAM = BiocParallel::SerialParam(),
  recalculate = FALSE, verbose = FALSE, ...)

Arguments

Y
The data (genes in rows, samples in columns). Currently implemented only for SummarizedExperiment.

...
Passthrough arguments to zinbwave::zinbwave().

K
integer. Number of latent factors.

epsilon
nonnegative scalar. Regularization parameter.

BPPARAM
BiocParallel param. We recommend one of the following:

BiocParallel::bpparam().

BiocParallel::SerialParam().

BiocParallel::MulticoreParam().

recalculate
logical. Force recalculation of weights.

verbose
logical. Run zinbwave in verbose model (for debugging).

Value

Modified object (S4 class must extend SingleCellExperiment), with weights added to assays().

Note

zinbwave defaults to using BiocParallel::bpparam() to register the number of cores to use (BPPARAM argument), but we've found that this works inconsistently across installations. Currently, we recommend using BiocParallel::SerialParam() by default, which will run in serial, using a single core. On Linux or macOS, BiocParallel::MulticoreParam() should work to run with multiple cores.

See also

Examples

# Example using 200 non-zero genes
Y <- sce_small
Y <- bcbioSingleCell::filterCells(Y)

#> Parameters:
#>   - >= 0 UMIs per cell
#>   - <= Inf UMIs per cell
#>   - >= 0 genes per cell
#>   - <= Inf genes per cell
#>   - >= 0 novelty score
#>   - <= 1 mitochondrial abundance
#>   - == Inf cells per sample
#>   - >= 1 cells per gene
#> ================================================================================
#> Cells:
#>      100 cells | prefilter
#>      100 cells | minUMIs >= 0
#>      100 cells | maxUMIs <= Inf
#>      100 cells | minGenes >= 0
#>      100 cells | maxGenes <= Inf
#>      100 cells | minNovelty <= 0
#>      100 cells | maxMitoRatio <= 1
#>      100 cells | nCells == Inf
#> 
#> group1 group2 
#>     49     51 
#> ================================================================================
#> Genes:
#>     1201 genes | prefilter
#>     1201 genes | minCellsPerGene >= 1
#> ================================================================================
#> Summary:
#>   - 100 of 100 cells passed filtering (100%)
#>   - 1201 of 1201 genes passed filtering (100%)

Y <- Y[seq_len(200L), ]
zinb <- suppressMessages(runZinbwave(
    Y = Y,
    BPPARAM = BiocParallel::SerialParam(),
    recalculate = TRUE
))
print(zinb)

#> class: SingleCellExperiment 
#> dim: 200 100 
#> metadata(4): filterCells filterGenes filterParams filterSummary
#> assays(5): counts logcounts scaleData normalizedValues weights
#> rownames(200): ENSG00000000938 ENSG00000001460 ... ENSG00000080845
#>   ENSG00000080910
#> rowData names(7): broadClass description ... geneName seqCoordSystem
#> colnames(100): cell1 cell10 ... cell98 cell99
#> colData names(16): nGene nUMI ... res1.2 ident
#> reducedDimNames(3): PCA TSNE UMAP
#> spikeNames(0):

assayNames(zinb)

#> [1] "counts"           "logcounts"        "scaleData"        "normalizedValues"
#> [5] "weights"