Run ZINB-WaVE

Run ZINB-WaVE

runZinbwave(Y, ...)

# S4 method for SingleCellExperiment
runZinbwave(
  Y,
  K = 0L,
  epsilon = 1e+12,
  BPPARAM = BiocParallel::bpparam(),
  recalculate = FALSE,
  verbose = FALSE
)

Arguments

Y
The data (genes in rows, samples in columns).

...
Additional arguments.

K
integer. Number of latent factors.

epsilon
nonnegative scalar. Regularization parameter.

BPPARAM
bpparamClass. BiocParallel parameter to specify the desired processor configuration.
We recommend using one of the following:

bpparam.

SerialParam.

MulticoreParam.

recalculate
logical. Force recalculation of weights.

verbose
logical. Run zinbwave in verbose mode (for debugging).

Value

Modified object. The dimensionality reduced matrix is stored in the reducedDims slot, and optionally normalized values and residuals are added in the list of assays.

Details

zinbwave will calculate normalizedValues and weights matrices, which will be defined in the assays slot. If these values have already been calculated, the object will be returned unmodified, unless recalculate = TRUE.

Note

Updated 2020-07-23.

Parallelization

zinbwave() defaults to using BiocParallel::bpparam to register the number of cores to use (bpparam argument), but we've found that this works inconsistently across installations. Currently, we recommend using BiocParallel::SerialParam by default, which will run in serial, using a single core. On Linux or macOS, BiocParallel::MulticoreParam should work to run with multiple cores.

See also

Examples

if (requireNamespace("zinbwave", quietly = TRUE)) {
    data(SingleCellExperiment, package = "AcidTest")
    Y <- SingleCellExperiment
    Y <- nonzeroRowsAndCols(Y)
    Y <- runZinbwave(Y)
}

#> ℹ Filtered zero count rows and columns:
#> - 483 / 500 rows (97%)
#> - 100 / 100 columns (100%)

#> → Running zinbwave.

#> ℹ BiocParallel param registered.
#> class: MulticoreParam
#>   bpisup: FALSE; bpnworkers: 6; bptasks: 0; bpjobname: BPJOB
#>   bplog: FALSE; bpthreshold: INFO; bpstopOnError: TRUE
#>   bpRNGseed: ; bptimeout: 2592000; bpprogressbar: TRUE
#>   bpexportglobals: TRUE
#>   bplogdir: NA
#>   bpresultdir: NA
#>   cluster type: FORK

#> → `zinbFit()`: Fitting a ZINB regression model.

#> CPU time used:
#>    user  system elapsed 
#>   9.465   5.804   3.686 

#> Object of class ZinbModel. 100 samples; 483 genes. 1 sample-level covariate(s)
#> (mu); 1 sample-level covariate(s) (pi); 1 gene-level covariate(s) (mu); 1
#> gene-level covariate(s) (pi); 0 latent factor(s).

#> → `zinbwave()`: Performing dimensionality reduction using a ZINB regression model with gene and cell-level covariates.

#> CPU time used: user system elapsed 0.004 0.000 0.004